refactor notification logic
