--- test/cypress/fixtures/tableNotifications.json
@@ -1,77 +1,0 @@
-[
-  {
-    "mark_as_read": "Mark as read",
-    "mark_as_unread": "Mark as unread",
-    "id": 1,
-    "level": "info",
-    "unread": true,
-    "deleted": false,
-    "verb": "New message",
-    "description": "You have a new message from your coordinator.",
-    "timestamp": "2023-05-15T10:30:00Z",
-    "data": {
-      "url": "",
-      "icon": "mail"
-    }
-  },
-  {
-    "mark_as_read": "Mark as read",
-    "mark_as_unread": "Mark as unread",
-    "id": 2,
-    "level": "success",
-    "unread": false,
-    "deleted": false,
-    "verb": "Challenge completed",
-    "description": "Congratulations! You've completed the weekly challenge.",
-    "timestamp": "2023-05-14T18:45:00Z",
-    "data": {
-      "url": "/challenges/weekly",
-      "icon": "mdi-trophy"
-    }
-  },
-  {
-    "mark_as_read": "Mark as read",
-    "mark_as_unread": "Mark as unread",
-    "id": 3,
-    "level": "warning",
-    "unread": true,
-    "deleted": false,
-    "verb": "Profile update required",
-    "description": "Please update your profile information.",
-    "timestamp": "2023-05-13T09:15:00Z",
-    "data": {
-      "url": "/profile",
-      "icon": "person"
-    }
-  },
-  {
-    "mark_as_read": "Mark as read",
-    "mark_as_unread": "Mark as unread",
-    "id": 4,
-    "level": "error",
-    "unread": false,
-    "deleted": false,
-    "verb": "Payment failed",
-    "description": "Your recent payment for the challenge registration has failed.",
-    "timestamp": "2023-05-12T14:20:00Z",
-    "data": {
-      "url": "/payments",
-      "icon": "warning"
-    }
-  },
-  {
-    "mark_as_read": "Mark as read",
-    "mark_as_unread": "Mark as unread",
-    "id": 5,
-    "level": "info",
-    "unread": true,
-    "deleted": false,
-    "verb": "New team member",
-    "description": "A new member has joined your team.",
-    "timestamp": "2023-05-11T11:00:00Z",
-    "data": {
-      "url": "",
-      "icon": "group_add"
-    }
-  }
-]
